feat(frontend): :building_construction: Split pages for overlay and app
Remove old homepage, have two new pages: app and overlay, adds the mouse listener for overlay
