+++ src/routes/app/+layout.svelte
@@ -1,5 +1,6 @@
 <script>
     let { children } = $props();
+    import "../../app.css";
 </script>
 
 {@render children()}
